split imap class
