+++ FlowCrypt/Common/Imap/Imap+folders.swift
@@ -18,6 +18,7 @@ protocol FoldersProvider {
 }
 
 extension Imap: FoldersProvider {
+
     func fetchFolders() -> Promise<FoldersContext> {
         return Promise { [weak self] resolve, reject in
             self?.getImapSess()?
@@ -37,4 +38,5 @@ extension Imap: FoldersProvider {
                 }
         }
     }
+
 }
